requirements.txt
Werkzeug==0.14.1
Flask_SQLAlchemy==2.3.2
Flask_RESTful==0.3.6
Flask==1.0.2
PyJWT==1.6.4
